requirements.txt
fastapi
uvicorn
motor  # For MongoDB async driver
pydantic
python-dotenv
